Keep everything grouped under the `org.asciidctor.asciidoclet` group and package.
