+++ src/main/java/org/asciidoctor/asciidoclet/JavadocParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoclet.asciidoclet;
+package org.asciidoctor.asciidoclet;
 
 import java.util.ArrayList;
 import java.util.List;
